smdk4412-common: a dirty workaround to MTP for buggy hosts
The currently used FunctionFS implementation seems to
have some issues with Samsung MTP driver initialization
on Windows 10 OS.

Add service that restarts the USB config, also do this twice
to ensure that the driver is initialized.

This is a cruel hack, but I don't know what else can be done.

Change-Id: I25fcfc7c7ed91a0e58b4b3a3b985027d145b254a